# internal/catalog/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Syscat inventory API",
    "version": "1",
    "description": "Create and search inventory, read records and photographs, attach photos, and PATCH current descriptions, locations, captions, and groups. Writes currently use trusted-network access without authentication, cookies, or CSRF tokens. Edits and attachments require the current asset revision; creation starts a new record at revision 1. URLs are root-relative."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/assets": {
      "get": {
        "operationId": "listAssets",
        "summary": "Search or list inventory",
        "description": "Newest inventory IDs first. Defaults to active records. Unknown, repeated, or invalid query parameters return 400. Original intake is omitted from listing items; fetch the individual record for it. Previous and next links are omitted when unavailable. Use view=summary for compact catalog review without description bodies, locations, timestamps, or photo metadata; follow api_url for full context. Pagination retains the selected view. Omitting view preserves the existing full response. Use field to restrict q to titles, descriptions, locations, captions, or IDs. Pagination preserves explicit field selection.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive substring search in description, location, and current photo captions, or exact numeric ID (with or without zero padding).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Select archived records instead of active records. Empty is equivalent to false.",
            "schema": {
              "type": "string",
              "enum": [
                "",
                "0",
                "1",
                "false",
                "true"
              ],
              "default": "0"
            }
          },
          {
            "name": "view",
            "in": "query",
            "description": "Response projection. Summary records contain identity, display title, revision, photo count, archive state, and record links.",
            "schema": {
              "type": "string",
              "enum": [
                "full",
                "summary"
              ],
              "default": "full"
            }
          },
          {
            "name": "field",
            "in": "query",
            "description": "Restrict q to one current field. all preserves broad search; title searches the complete first description line (before display truncation), description includes that line and the body, caption searches any current photo, and id matches an exact positive numeric inventory ID with optional zero padding. Empty q lists the selected archive set.",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "title",
                "description",
                "location",
                "caption",
                "id"
              ],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated inventory",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AssetList"
                    },
                    {
                      "$ref": "#/components/schemas/AssetSummaryList"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAsset",
        "summary": "Create an asset with original intake and photos",
        "description": "Trusted-network multipart creation without cookies or CSRF tokens. Requires a description and exactly one overview photo. Location, overview caption, detail photos, detail captions, and groups are optional. Uses the shared image metadata stripping/thumbnail pipeline and captures original intake. Returns revision 1. Optional Idempotency-Key identifies a logical creation: first successful creation wins; replay returns that record in its current state with 200, ignoring replacement content after basic request validation. Keys persist across restarts and archiving. Use a new key for each different creation. Without a key, every successful request creates a new asset. A failed creation does not reserve a key. Unknown fields and query parameters are rejected.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Optional persistent retry key. 1\u2013128 ASCII letters, digits, dots, underscores, or hyphens. Must appear once; case sensitive. Do not reuse for another logical creation.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._-]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              },
              "encoding": {
                "photos": {
                  "contentType": "image/jpeg, image/png, image/gif"
                },
                "caption_photos": {
                  "style": "form",
                  "explode": true
                },
                "group_photos": {
                  "style": "form",
                  "explode": true
                },
                "overview": {
                  "contentType": "image/jpeg, image/png, image/gif"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created asset including permanent inventory ID, stable photo IDs/links, original intake, and revision 1.",
            "headers": {
              "Location": {
                "description": "Root-relative URL of the updated asset.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "400": {
            "description": "Missing description/overview, malformed multipart request, invalid idempotency key or metadata, unknown/duplicate fields, or invalid image.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds configured max-upload-mib (default 256 MiB), or a photo exceeds 12 MiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Use multipart/form-data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Idempotent replay: current record created by this key, including later edits or archived status. No duplicate images or records.",
            "headers": {
              "Location": {
                "description": "Root-relative URL of the updated asset.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{id}": {
      "get": {
        "operationId": "getAsset",
        "summary": "Read one inventory record",
        "description": "Includes original intake. Archived records are readable by ID. Accepts numeric IDs with or without zero padding.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "Positive inventory ID within signed 64-bit range."
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchAsset",
        "summary": "Edit asset text",
        "description": "Trusted-network write; no credentials or browser session required. Query parameters are not accepted. Photo PATCH returns the full updated asset. IDs, intake, timestamps, files, order, and archive state cannot be directly overwritten.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated full asset with its incremented revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, unsupported fields, or invalid edits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Asset or attached photo not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stale revision or archived asset. Read the latest asset before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 128 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assets/{id}/photos/{photo_id}": {
      "patch": {
        "operationId": "patchPhoto",
        "summary": "Edit a photo caption or group",
        "description": "Trusted-network write; no credentials or browser session required. Query parameters are not accepted. Photo PATCH returns the full updated asset. IDs, intake, timestamps, files, order, and archive state cannot be directly overwritten.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the stable id from a current photo, or follow its api_url."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated full asset with its incremented revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, unsupported fields, or invalid edits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Asset or attached photo not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stale revision or archived asset. Read the latest asset before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 128 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{id}/photos": {
      "post": {
        "operationId": "uploadAssetPhotos",
        "summary": "Attach photos to an existing asset",
        "description": "Trusted-network multipart upload without cookies or CSRF tokens. Requires the current asset revision. One or more photos are saved as one revision, with optional captions and groups aligned by file order. Preserves existing overview, metadata, and original intake. Reuses image validation, metadata stripping, and thumbnail generation. On any failure the entire batch is rejected and its generated files removed. Concurrent changes return 409; reread and reconcile before retrying. No query parameters accepted.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PhotoUpload"
              },
              "encoding": {
                "photos": {
                  "contentType": "image/jpeg, image/png, image/gif"
                },
                "caption_photos": {
                  "style": "form",
                  "explode": true
                },
                "group_photos": {
                  "style": "form",
                  "explode": true
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Full updated asset with stable photo IDs and new revision. New files are assigned detail roles; the existing overview is retained. For a legacy asset with no photos, its first attachment becomes the ungrouped overview.",
            "headers": {
              "Location": {
                "description": "Root-relative URL of the updated asset.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "400": {
            "description": "Malformed multipart request, missing or invalid revision, invalid metadata, or unsupported/invalid image.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Asset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stale revision or archived asset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds configured max-upload-mib (default 256 MiB), or a photo exceeds 12 MiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Use multipart/form-data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Asset"
          },
          {
            "type": "object",
            "required": [
              "original_intake"
            ]
          }
        ]
      },
      "AssetList": {
        "type": "object",
        "required": [
          "assets",
          "total",
          "page",
          "page_size",
          "query",
          "archived"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "query": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "previous": {
            "type": "string",
            "format": "uri-reference"
          },
          "next": {
            "type": "string",
            "format": "uri-reference"
          },
          "field": {
            "type": "string",
            "enum": [
              "all",
              "title",
              "description",
              "location",
              "caption",
              "id"
            ],
            "description": "Selected search field. Present only when field is supplied in the request."
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "invalid_id",
                  "not_found",
                  "method_not_allowed",
                  "internal_error",
                  "conflict",
                  "unsupported_media_type",
                  "request_too_large",
                  "unauthorized",
                  "forbidden"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "AssetSummaryList": {
        "type": "object",
        "required": [
          "assets",
          "total",
          "page",
          "page_size",
          "query",
          "archived",
          "view"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "query": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "previous": {
            "type": "string",
            "format": "uri-reference"
          },
          "next": {
            "type": "string",
            "format": "uri-reference"
          },
          "view": {
            "type": "string",
            "enum": [
              "summary"
            ]
          },
          "field": {
            "type": "string",
            "enum": [
              "all",
              "title",
              "description",
              "location",
              "caption",
              "id"
            ],
            "description": "Selected search field. Present only when field is supplied in the request."
          }
        },
        "description": "Summary projection of a listing; pagination links retain view=summary and other filters."
      },
...
    }
  }
}